deprecate configuration.py
